moved states array out of form into constants folder
